Refactored models to ProgramInput and DualProgramInput objects and Output models with a .results attribute corresponding to the CalcType. Removed redundant docstrings.
