+++ pyproject.toml
@@ -16,6 +16,7 @@ pydantic = ">=1.7.4,!=1.8,!=1.8.1,<2.0.0"
 numpy = ">=1.20"
 toml = "^0.10.2"
 pyyaml = "^6.0"
+typing-extensions = "^4.7.1"
 
 [tool.poetry.group.dev.dependencies]
 black = "^23.3.0"
@@ -25,7 +26,6 @@ pre-commit = "^3.2.1"
 pytest-cov = "^4.0.0"
 ruff = "^0.0.260"
 isort = "^5.12.0"
-typing-extensions = "^4.5.0"
 qcelemental = "^0.25.1"
 types-toml = "^0.10.8.6"
 types-pyyaml = "^6.0.12.10"
