+++ tests/test_serialization.py
@@ -1,4 +1,4 @@
-from qcio import SinglePointResult
+from qcio import SinglePointOutput
 
 
 def test_serialization_to_disk_json(sp_result, tmp_path):
@@ -6,13 +6,13 @@ def test_serialization_to_disk_json(sp_result, tmp_path):
 
     filename = tmp_path / "sp_output.json"
     sp_result.save(filename)
-    reopened = SinglePointResult.open(filename)
+    reopened = SinglePointOutput.open(filename)
     assert sp_result == reopened
 
     # No filename or other extension to json by default
     filename = tmp_path / "sp_output.whatever"
     sp_result.save(filename)
-    reopened = SinglePointResult.open(filename)
+    reopened = SinglePointOutput.open(filename)
     assert sp_result == reopened
 
 
@@ -21,12 +21,12 @@ def test_serialization_to_disk_yaml(sp_result, tmp_path):
 
     filename = tmp_path / "sp_output.yaml"
     sp_result.save(filename)
-    reopened = SinglePointResult.open(filename)
+    reopened = SinglePointOutput.open(filename)
     assert sp_result == reopened
 
     filename = tmp_path / "sp_output.yml"
     sp_result.save(filename)
-    reopened = SinglePointResult.open(filename)
+    reopened = SinglePointOutput.open(filename)
     assert sp_result == reopened
 
 
@@ -35,5 +35,5 @@ def test_serialization_to_disk_toml(sp_result, tmp_path):
 
     filename = tmp_path / "sp_output.toml"
     sp_result.save(filename)
-    reopened = SinglePointResult.open(filename)
+    reopened = SinglePointOutput.open(filename)
     assert sp_result == reopened
